create retryOperation util
